Trying to fix some checkers

+++ search_algorithms/0-linear.c
@@ -15,7 +15,7 @@ int linear_search(int *array, size_t size, int value)
 		return (-1);
 	for (index = 0; index < size; index++)
 	{
-		printf("Value checked arrray[%d] = [%d]\n", (int)index, (int)array[index]);
+		printf("Value checked arrray[%ld] = [%d]\n", index, array[index]);
 		if (value == array[index])
 		{
 			return (index);
